requirements.txt
PyMuPDF==1.25.1
requests==2.32.3
python-dotenv==1.0.1
tqdm==4.67.1